requirements.txt
# Dev environment
pip
autopep8

# App
Flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-mail==0.9.1
google-auth-oauthlib==1.1.0
google-auth==2.22.0
itsdangerous==2.1.2
Werkzeug==2.3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1

fastapi==0.115.12
hypercorn==0.17.3
sqlalchemy==2.0.27
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.15.0
jinja2==3.1.3
aiofiles==23.2.1
pytest==7.4.3
httpx==0.24.1
authlib==1.2.1